requirements.txt
streamlit==1.29.0
openai==1.6.1
python-dotenv==1.0.0
streamlit-audiorecorder==0.0.5
